# internal/api/spec/api-spec.yaml
openapi: 3.0.0
info:
  title: Hyperboard API
  version: 0.1.0
paths:
...
  /api/v1/tags:
    get:
      operationId: getTags
      summary: List tags.
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          $ref: "#/components/responses/TagsResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/tags/{tag}:
    put:
      operationId: putTag
      summary: Create or replace a tag by name.
      parameters:
        - $ref: "#/components/parameters/tag"
      requestBody:
        $ref: "#/components/requestBodies/TagRequest"
      responses:
        "201":
          $ref: "#/components/responses/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    get:
      operationId: getTag
      summary: Get a tag by name.
      parameters:
        - $ref: "#/components/parameters/tag"
      responses:
        "200":
          $ref: "#/components/responses/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
    delete:
      operationId: deleteTag
      summary: Delete a tag. The tag will be removed from any posts and tag categories.
      parameters:
        - $ref: "#/components/parameters/tag"
      responses:
        "204":
          $ref: "#/components/responses/DeletedResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
  /api/v1/tags/{tag}/convert-to-alias:
    post:
      operationId: convertTagToAlias
      summary: Convert a tag to an alias of another tag. Re-tags all posts and merges aliases atomically.
      parameters:
        - $ref: "#/components/parameters/tag"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - target
              properties:
                target:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/TagResponse"
        "400":
          $ref: "#/components/responses/BadRequestResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsResponse"
        "500":
          $ref: "#/components/responses/InternalServerErrorResponse"
...
components:
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    Cursor:
      type: string
    Limit:
      type: integer
      minimum: 1
      maximum: 1024
      default: 24
...
  parameters:
...
    tag:
      name: tag
      in: path
      required: true
      schema:
        $ref: "types-spec.yaml#/components/schemas/TagName"
    cursor:
      name: cursor
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Cursor"
    limit:
      name: limit
      in: query
      required: false
      schema:
        $ref: "#/components/schemas/Limit"
...
  requestBodies:
    TagRequest:
      required: true
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/Tag"
...
  responses:
...
    DeletedResponse:
      description: Resource deleted successfully
    BadRequestResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnauthorizedResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ForbiddenResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFoundResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequestsResponse:
      headers:
        Retry-After:
          schema:
            type: integer
            minimum: 0
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalServerErrorResponse:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TagResponse:
      content:
        application/json:
          schema:
            $ref: "types-spec.yaml#/components/schemas/Tag"
    TagsResponse:
      content:
        application/json:
          schema:
            type: object
            properties:
              items:
                type: array
                items:
                  $ref: "types-spec.yaml#/components/schemas/Tag"
              cursor:
                $ref: "#/components/schemas/Cursor"
...